Implemented full creation of the FlightSection and Seat class
will now fully fill the double array of Seats contained within a
SeatClass, and check for available seats as well as book the first
available seat.  Haven't implemented the functions in the system manager
yet, but it should just be calling them and checking for any conditions
to not find them.