Added NullObjects, Cleared up the DataClasses, and added Air's Section!
- added NullTransition so that I could both return null Transitions
without making it wig out.
- Therefore, I put the Data storage class inside transition.  This way I
can initiailze transit as a NullTransition, and then use it to create
the DataClass for whatever is gonna be made!
- Also put a Data storage class inside Section, both are now just called
DataClass, with whatever class they are storing data for in front of
them when called, so: Transition.DataClass or Section.DataClass.
- I also have almost fully fleshed out the factories for creating Air's
stuff.  Not entirely there yet, but almost.