Added ship functionality, added test Client, improved MyDate methods
- Ship is now a full class, with the capability to book a ship if the
ship isn't already occupied for that time.
- All ships are initialized during the creation of a cruiseline, with
the added on Strings being counted as the extra ships to be added to the
line during creation.
- Test client is here! and it's about as boring as to be expected, but
it does do the tests that I want it to, cause I've got too much on my
plate to begin with to get used to making JUnit Tests for everything...