Worked on Everything Cruise, changes to abstracts, layout now in Air!
- Finished all of cruise factory, so section now will add a section to
the given transID assuming the transID is a ship (yeah I know a bit of a
stretch, I might change it) and always return a nullSection after adding
but print out an error message if it doesn't work.
- CruiseShip, CruiseTrip, CruiseSection, and CruiseLine should all but
fully implemented and functional.
- All abstract Transport classes now have either final by the field if
it's a packaged private, so it can only be initialized during the
constructor and never modified anywhere else, or are fully private with
getters if it's something that might be added too.
- Moved some fields such as seatingClass to abstracted fields.
- TransportClass is now the interface for both SeatClass, and the new
CabinClass enum specifically for cruises.  I considered having a boolean
on the cabin class for if it's deluxe or not, but settled on just having
4 classes in cabin class for family, couples and their deluxe versions.