Actually finished Everything Cruise, abstracted to TransportSeating
- Abstracted out Seating to TransportSeating, which now extends to Seat
and Cabin.
- Actually finished CruiseSection, so it now contains a
HashMap<Character, Cabin[]>, Characters for each deck of the Cruiseship,
and cabin's number as the index locations.  It also creates the entire
thing out of an int of cabins per deck, and a string, with each
character denoting a deck's character name (yeah if I get to it I might
do it differently and better but for now...)