Client-side work, toStrings rehauled, Options Galore, Admin awakens
- Done a ton of menu work, allowing for all sorts of nifty things to be
done.
- All Air and Cruise paraphanalia now have both a toString method that
returns a condensed form for printing to file, and a toViewingString
method that prints a readable form of the information.
- You can choose at the start of runtime whether to operate in air land
or cruise land, and pretty soon coming to theaters near you, will be the
option to switch back and forth between admin and customer mode!!!!!!
O.o o.O
- admin is now awake and beginning to stir.
- finally given up on doing the trains, just not enough time.