Cleaner Cruise toStrings, Cruise pre-fab, work in progress
- a lot of the Cruise toViewingStrings were still kinda messy and didn't
do it all that much justice of how it was set up, so I worked on
cleaning them up, and removed toStrings since I wasn't using them in
cruises, therefore they needed to be culled and removed from the parent
class.
- set up a pre-fabricated set of cruises in the client to test viewing
with, will remove it after I am done (I have put in lots of comments to
remind myself)
- I'm stopping in the middle of working out some way to properly attach
a cruise trip to it's accompanying ship, since I realized they weren't
actually really attached other than the ship is booked for a duration,
and the trip has a ship available during it's time (which is sorta good
encapsulation), however in printing out, I need to print what trips each
ship is scheduled for, so the cruiseline does need to know what go
together.