Finally finished cabin sections adding, ready to get outside evaluation!
- cabins weren't adding correctly to new Cruise trips, I figured out it
was because when I added a section to a ship, I forgot to add that
section to any trips that boat has already been booked for, so it could
work if I made the sections first and then added a transit to the boat,
but after fixing that, it now works either way!
- I believe this is satisfactorily working enough for me to take it to
my teacher and get some input on what I could do for using it as a
show-case project for possible jobs!  I will also create a reflection
over if I were to do this entire project again from square one, what
things I would do differently (planning, oh god I could have done so
much more and better planning), and what things I did well that I would
do again.